fix: prevent infinite bind mounts

+++ fastrpc/usr/libexec/fastrpc/setup-dsp.sh
@@ -14,9 +14,11 @@ case "$MAC" in
     "Radxa Dragon Q8B")
         ln -sf /usr/share/qcom/sc8280xp/radxa/dragon-q8b/dsp /usr/lib/dsp
         SOC=498
-        mkdir -p /tmp/fake-soc0
-        echo "$SOC" > /tmp/fake-soc0/soc_id
-        mount --bind /tmp/fake-soc0/soc_id /sys/devices/soc0/soc_id
+        if [[ ! -f /tmp/fake-soc0/soc_id ]]; then
+            mkdir -p /tmp/fake-soc0
+            echo "$SOC" > /tmp/fake-soc0/soc_id
+            mount --bind /tmp/fake-soc0/soc_id /sys/devices/soc0/soc_id
+        fi
         ;;
     "Radxa AIRbox Q900")
         ln -sf /usr/share/qcom/qcs9075/radxa/airbox-q900/dsp /usr/lib/dsp
